+++ src/main/java/com/example/electric_storage/api/resistor/ResistorResponse.java
@@ -3,4 +3,8 @@ package com.example.electric_storage.api.resistor;
 import com.example.electric_storage.infrastructure.common.ValueMultiplier;
 
 public record ResistorResponse(
-    String uniqueId, int resistance, ValueMultiplier multiplier, float tolerance, float power) {}
+    String uniqueId,
+    double resistance,
+    ValueMultiplier multiplier,
+    double tolerance,
+    double power) {}
